refactors all internal state types into a single module
